Standardize JWT handling
internal
--------
Rename the shared secret from `jwt-secret.secret` to `secrets.jwt-secret`.

Move account_*.yaml configs into `accounts/` where they belong.

content
-------
Make the writer use the new name for the JWT secret.

accounts
--------
Use the `secrets.jwt-secret` Kubernetes secret so all components share the
same JWT secret, which is kept hidden from git.

Put the JWT secret in the `JWT_SECRET` environment variable instead of
jwt_token, since it is a secret not a token (and to be consistent).

Give `saltRounds` parameter a name to avoid magical anonymous constants.

Make `signToken` return a `Promise` to avoid blocking in synchronous calls.

Use the async bcrypt `hash` function to avoid blocking in synchronous calls.

Make the `/register` and `/login` endpoints return the token in the response.
JWT are not cookies, they're bearer tokens.

Return responses with the proper status codes and useful error messages
instead of throwing exceptions on bad credentials.

Homoginize response JSON to always include a "status" key, with an
"error" message if errors are encountered or a "token" otherwise.